requirements.production.txt
# ============================================================================
# EVOLVE TRADING PLATFORM - PRODUCTION REQUIREMENTS
# ============================================================================
# Optimized for production deployment with security and performance focus

# ============================================================================
# CORE DEPENDENCIES
# ============================================================================
# Essential packages for the trading system
numpy>=1.21.0,<2.0.0
pandas>=1.5.0,<2.0.0
scikit-learn>=1.1.0,<2.0.0
scipy>=1.7.0,<1.10.0
matplotlib>=3.5.0,<4.0.0
plotly>=5.10.0,<6.0.0
streamlit>=1.28.0,<2.0.0
pydantic>=1.8.0,<2.0.0
requests>=2.28.0,<3.0.0
websockets>=10.0,<11.0.0

# ============================================================================
# FORECASTING MODELS
# ============================================================================
# Machine learning and deep learning models
torch>=1.12.0,<2.0.0
torchvision>=0.13.0,<1.0.0
torchaudio>=0.12.0,<1.0.0
pytorch-lightning>=1.8.0,<2.0.0
xgboost>=1.6.0,<2.0.0
catboost>=1.1.0,<2.0.0
prophet>=1.1.0,<2.0.0
statsmodels>=0.13.0,<1.0.0
arch>=5.3.0,<6.0.0

# ============================================================================
# VISUALIZATION
# ============================================================================
# Charting and visualization libraries
seaborn>=0.11.0,<1.0.0
networkx>=2.6.0,<3.0.0
mplfinance>=0.12.0,<1.0.0

# ============================================================================
# STRATEGY & TRADING
# ============================================================================
# Trading-specific libraries and data providers
yfinance>=0.1.87,<1.0.0
alpha-vantage>=2.3.1,<3.0.0
pandas-ta>=0.3.0,<1.0.0
finta>=1.3,<2.0.0
empyrical>=0.5.5,<1.0.0

# ============================================================================
# MONITORING & PRODUCTION
# ============================================================================
# Logging, monitoring, and production tools
loguru>=0.6.0,<1.0.0
python-json-logger>=2.0.0,<3.0.0
sentry-sdk>=1.3.0,<2.0.0
psutil>=5.8.0,<6.0.0
aiohttp>=3.8.0,<4.0.0
apscheduler>=3.9.0,<4.0.0

# ============================================================================
# DATABASE & STORAGE
# ============================================================================
# Database connections and data storage
pymongo>=3.12.0,<4.0.0
redis>=4.3.0,<5.0.0
sqlalchemy>=1.4.0,<2.0.0
boto3>=1.26.0,<2.0.0
cloudpickle>=2.0.0,<3.0.0

# ============================================================================
# SECURITY
# ============================================================================
# Security and authentication
bcrypt>=3.2.0,<4.0.0
cryptography>=3.4.0,<4.0.0
passlib>=1.7.4,<2.0.0
python-jose>=3.3.0,<4.0.0

# ============================================================================
# REPORT GENERATION
# ============================================================================
# Report and PDF generation
jinja2>=3.0.0,<4.0.0
reportlab>=3.6.0,<4.0.0
fpdf>=1.7.0,<2.0.0

# ============================================================================
# AI & LLM
# ============================================================================
# AI and language model integration
openai>=1.0.0,<2.0.0

# ============================================================================
# OPTIMIZATION
# ============================================================================
# Optimization libraries
optuna>=3.0.0,<4.0.0

# ============================================================================
# DATA PROCESSING & ANALYSIS
# ============================================================================
python-dotenv>=0.19.0,<1.0.0
joblib>=1.1.0,<2.0.0
tqdm>=4.64.0,<5.0.0

# ============================================================================
# NLP & AI
# ============================================================================
nltk>=3.7,<4.0.0
spacy>=3.4.0,<4.0.0

# ============================================================================
# ADVANCED FEATURES
# ============================================================================
cvxpy>=1.2.0,<2.0.0
cvxopt>=1.3.0,<2.0.0
gymnasium>=0.26.0,<1.0.0
stable-baselines3>=1.6.0,<2.0.0
newsapi-python>=0.2.6,<1.0.0
fredapi>=0.4.0,<1.0.0
wbdata>=0.3.0,<1.0.0

# ============================================================================
# PRODUCTION MONITORING
# ============================================================================
# Production monitoring and observability
prometheus-client>=0.11.0,<1.0.0
grafana-api>=1.0.0,<2.0.0

# ============================================================================
# CONTAINERIZATION
# ============================================================================
# Container and deployment tools
gunicorn>=20.1.0,<21.0.0
uvicorn>=0.15.0,<1.0.0

# ============================================================================
# TESTING (MINIMAL FOR PRODUCTION)
# ============================================================================
# Basic testing for production validation
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.15.0,<1.0.0 